[#22] Add flux abstractions and extract sources.
New classes were created for face extrapolation, inviscid flux, and
viscous flux calculations. The fluxes compute the source terms and thus
inherit from the RightHandSide base class. Fluxes are also associated
with a face_extrapolator routine, which is in turn associated with a
number of ghost layers required by the extrapolation routine.

The overall speed of the program appears to have been reduced, and one
obvious factor is the repeated calculations within ViscousFlux where
transport properties are now computed for both sides of each face, even
though the right side of face i is the same as the left side of face i-1
for most faces when using linear extrapolation. There are multiple ways
to improve this, and one which is generalizable is to average the left
and right primitive values before computing the transport properties,
rather than averaging the transport properties from the left and right.

Rewrote the conservative <-> primitive transforms to simplify several
aspects. The conservative variables are kept within a single array while
the primitive variables are kept within a FluidState dataclass.