Generalize source terms (#28)
* [#22] Move SkinFriction into a new `models` directory.

* [#22] Extracted boundary layer source terms.

* [#22] Change FluidState shape to a tuple.

This change is in preparation for supporting shapes with more than one
dimension, which is currently needed for the left/right face states.

* [#22] Add flux abstractions and extract sources.

New classes were created for face extrapolation, inviscid flux, and
viscous flux calculations. The fluxes compute the source terms and thus
inherit from the RightHandSide base class. Fluxes are also associated
with a face_extrapolator routine, which is in turn associated with a
number of ghost layers required by the extrapolation routine.

The overall speed of the program appears to have been reduced, and one
obvious factor is the repeated calculations within ViscousFlux where
transport properties are now computed for both sides of each face, even
though the right side of face i is the same as the left side of face i-1
for most faces when using linear extrapolation. There are multiple ways
to improve this, and one which is generalizable is to average the left
and right primitive values before computing the transport properties,
rather than averaging the transport properties from the left and right.

Rewrote the conservative <-> primitive transforms to simplify several
aspects. The conservative variables are kept within a single array while
the primitive variables are kept within a FluidState dataclass.

* [#22] Formatting fixes.

* [#22] Add RightHandSide base class.

Forgot to add the base class for RightHandSide (i.e. time derivatives)
to an earlier commit.

* [#22] Move primitive <-> conservative transform into FluidPhysics.

* [#22] Move Array out of system/base.py

The Array type hint was causing a circular import.

* [#22] Standardize source terms around conservative variables.

RightHandSide evaluations are now by default based on a conservative
variable vector input, with gamma_star provided for calculating
primitive variables internally if needed.

InviscidFlux and ViscousFlux now perform face extrapolation internally,
followed by calling apply_boundary_conditions.

* [#20, #22] Extract geometry into its own class.

* Update more variables for case convention consistency

* Formatting (had to set up pre-commit)

* [#20, #22] Additional geometry-related updates.

* [#22] Initialize mass_fractions in laminar_flame example.

---------

Co-authored-by: Matthew Bonanni <mbonanni001@gmail.com>