[#10] Turn JICModel into RightHandSide.
Currently both get_chemical_sources and the former get_injector_sources
have adopted the signature of a RightHandSide.source term. We plan to
move the chemical source terms out of the JICModel later.

Note that advance_chemistry_FPV now lacks the internal energy
adjustment. This is being reworked under Issue #31, so we'll leave it
for now to maintain the source term structure.