[#23] Extract boundary conditions and change ghost layers.
Moving away from using "mt" in favor of "n_ghost_layers," and from
resizing the domain in favor of indexing into it with "idx_cells."

Made several typing-related updates along the way. Currently using
"assert ____ is not None" statements to type-narrow, but it is
cumbersome when checking that several elements of FluidState have been
set. I've looked for better ways, and tried creating a function like
"assert state.is_initialized()", but that did not successfully narrow
the types.