[#20] Decouple geometry from source terms.
In addition to pulling the source terms into their own AreaChange class,
the Geometry class was made more general by defining only a cross-
sectional area and perimeter. New subclasses for cylindrical and
box-like domains were created. The relationship between
height/width/diameters and the area, perimeter, hydraulic diameter, and
characteristic length were made more consistent, and all are treated as
functions of space and time for consistency. It is now possible to
initialize these with scalar values, arrays to be interpolated, or
direct callable functions.