Create class for the computational domain. (#48)
This update makes several improvements to the Geometry class and related infrastructure, including:

- Separating the source terms into their own RightHandSide subclass, `AreaChange`.
- Generalizing the cross-sectional area and perimeter inputs, and the hydraulic diameter and characteristic length derived properties.
- Adding subclasses for more specific flow geometries (`Box`, `Cylinder`, `AsymmetricBox`).
- Standardize handling of constant and linearly-interpolated values for the area and perimeter.
- Adding "regions" dictionary to facilitate tagging ranges of x-values within the domain.
- Separating the source terms into their own `RightHandSide` subclass.

Closes #20.


Detailed description below:

* [#20] Decouple geometry from source terms.

In addition to pulling the source terms into their own AreaChange class,
the Geometry class was made more general by defining only a cross-
sectional area and perimeter. New subclasses for cylindrical and
box-like domains were created. The relationship between
height/width/diameters and the area, perimeter, hydraulic diameter, and
characteristic length were made more consistent, and all are treated as
functions of space and time for consistency. It is now possible to
initialize these with scalar values, arrays to be interpolated, or
direct callable functions.

* [#20] Move AreaChange to its own module.

The area change related source terms have been moved from geometry to
models/area_change.py. This commit includes a couple minor formatting
and typing fixes.

* [#20] Speed up fast source term evaluation.

This change vectorizes the source_fast routine and adds a banded
Jacobian to the ODE solver. This change was manually cherry-picked from
issue branch 27, and needs to be verified.

Co-authored-by: nlaing3 <nlaing3@stanford.edu>

* [#20] Add "region" tags for geometric ranges.

Co-authored-by: nlaing3 <nlaing3@stanford.edu>

* [#20] Simplify geometric function initialization.

* [#20] Add asymmetric version of Box geometry.

This geometry allows the upper and lower walls of a box type geometry to
be specified independently.

Co-authored-by: nlaing3 <nlaing3@stanford.edu>

* [#20] Allow defining geometry with (x, y) tuples.

This update also simplifies the typing with a new type alias.

* [#20] Add geometry initialization function.

This function picks the Geometry [sub]class to initialize based on the
provided input parameters. It helps simplify the Combustor class, which
now only needs to pass kwargs onto initialize_geometry.

* [#20] Make automatic dlnA_dx work with (x, y) tuples.

* [#20] Fix TypeAlias import.

---------

Co-authored-by: nlaing3 <nlaing3@stanford.edu>