[#36] WIP: Update viscous flux terms
Adding the diffusive species enthalpy and viscous heating flux terms
to the viscous flux calculation. Additional modifications are still
required to support FPV approaches, as FPVTable currently lacks a species
enthalpy calculation and the species energy diffusive flux needs to be
mapped onto the flamelet scalar diffusive flux. The laminar flame case
currently blows up in the downstream region, and it's not clear yet if
that's due to a mistake in the formulation or something else.

Reworked the viscous flux to perform face-averaging first instead of
averaging the derived properties from either side. Also added a
correction velocity to the species mass diffusive fluxes to maintain
mass continuity.