[#36] Add chemical energy to transport (#43)
This adds the chemical energy to the energy transport equation, switching to a total energy formulation from the original sensible + kinetic. Additionally updates the double flux approach to include the energy term, "e0_star," to be consistent with the literature.

Detailed commit log below:

* [#36] Start conversion - need to look at interaction of FPV and double flux

* [#36] Finish implementing total energy formulation.

This update makes the use of gamma vs. gamma_star more explicit, adds
e0_star, and updates the fluxes and test cases accordingly.
Additionally, the conservative_to_primitive function was generalized to
work for both FRC and FPV modes, and could thus be moved to the
FluidPhysics base class.

* [#36] Add double-flux get_internal_energy.

* [#36] WIP: Update viscous flux terms

Adding the diffusive species enthalpy and viscous heating flux terms
to the viscous flux calculation. Additional modifications are still
required to support FPV approaches, as FPVTable currently lacks a species
enthalpy calculation and the species energy diffusive flux needs to be
mapped onto the flamelet scalar diffusive flux. The laminar flame case
currently blows up in the downstream region, and it's not clear yet if
that's due to a mistake in the formulation or something else.

Reworked the viscous flux to perform face-averaging first instead of
averaging the derived properties from either side. Also added a
correction velocity to the species mass diffusive fluxes to maintain
mass continuity.

* [#36] Add species enthalpy calculation for ThermoTable.

* [#36] Fix Z-calculation error in JICModel.

* [#36] Update JIC case with new geometry.

Also fixed a visualization bug due to `scalar_names` not having been
updated alongside the species transport update.

---------

Co-authored-by: Tim E. Dawson <tim.dawson@cfd-research.com>