[#55] Check inviscid flux against area change terms.
In theory, for an isentropic flow the inviscid flux source term should
be equal and opposite to the area change source term. Numerical error
due to, e.g., finite difference approximations will prevent this from
matching exactly.

Currently there is a significant difference near the boundaries and
right at the throat for a choked nozzle problem.