[#21] Add new operator-split approaches and RK4.
A more general base class for operator-splitting was added, along with
new versions of Strang and Lie splitting which utilize it and two
additional splitting approaches. Also added a more standard Runge-Kutta
base class and an implementation of RK4 which uses it, and some initial
documentation for some of the approaches.