[#70] Partial cleanup of Combustor signature.
Removed `kwargs` from `Combustor` and moved all keyword arguments with
their default values into the signature along with the comments
describing them. Boundary conditions must now be specified (no default
value, but can be a `BCInput` dictionary), and geometry must be passed
in as an initialized `Geometry` object.

This also enabled the removal of several attributes which are not used
outside of `__init__`, such as `inviscid_face_extrapolator` and
`include_boundary_layer`.

Closes #70