[#71] Add basic restart and slicing to FluidState.
This enables a basic read/write capability to the `FluidState` object
which can be used by the `InitializeRestart` initializer. The groupname
is currently treated as the iteration number, though this should be
decoupled in the future. A fully-featured restart capability is planned
which would also write the current iteration, time, and settings used
for the simulation to the restart file.

Also enabled slicing syntax, such that a subset of the FluidState array
is returned using syntax like `state[3:10]`.